moved workers to functions for roast-me
