--- cloudflare/api-server/test/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"compilerOptions": {
-		"types": ["@cloudflare/workers-types/experimental", "@cloudflare/vitest-pool-workers"]
-	},
-	"include": ["./**/*.ts", "../src/env.d.ts"],
-	"exclude": []
-}
