--- cloudflare/api-server/worker-configuration.d.ts
@@ -1,4 +1,0 @@
-// Generated by Wrangler
-// After adding bindings to `wrangler.toml`, regenerate this interface via `npm run cf-typegen`
-interface Env {
-}
